requirements.txt
streamlit
faiss-cpu
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu